Move `custom_process` to utils
